edited popupwindow test
