+++ TestRailFinalTask/TestRailFinalTask/Tests/GUI/ProjectTest.cs
@@ -39,20 +39,5 @@ namespace TestRailFinalTask.Tests.GUI
 
             Assert.That(dashboardPage.CheckAddedProjectMessageError, Is.EqualTo(expectedMessage));
         }
-
-        [Test]
-        public void PopUpWindow()
-        {
-            string email = "dashapr@mail.ru";
-            string password = "!Qwertyu123";
-
-            var dashboardPagePopUp = new LoginPage(Driver, true)
-                .SuccessfulLogin(email, password)
-                .CheckPopUpWindowAppeared();
-               
-            var expectedMessage = "Upgrade your account to start streamlining your testing!";
-
-            Assert.That(dashboardPagePopUp, Is.EqualTo(expectedMessage));
-        }
     }
 }
